pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rl-benchmark"
version = "0.1.0"
description = "Reproducible reinforcement learning benchmark for REINFORCE, A2C, and PPO."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "gymnasium[box2d]>=1.1,<1.3",
  "kaleido>=1.0",
  "numpy>=2.1",
  "pandas>=2.2",
  "plotly>=6.0",
  "PyYAML>=6.0",
  "streamlit>=1.44",
  "torch>=2.6",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3",
  "ruff>=0.11",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
